create trash controller, softDelete

+++ controllers/usersController.js
@@ -29,16 +29,14 @@ class UserController {
      * @returns res.json(data)
      */
     async show(req, res) {
+
         const keys = Object.keys(req.query);
 
         if (!keys[0]) {
-            res.status(404).json('Insira um valor para pesquisa!');
+           return res.status(404).json('Insira um valor para pesquisa!');
         }
 
         const data = await db.User.findOne({ where: { [keys[0]]: req.query[keys[0]] } });
-        if (!data) {
-            res.status(404).json('Usuário não encontrado!');
-        }
         res.status(200).json(data);
     }
 
@@ -49,6 +47,10 @@ class UserController {
      * @returns res.json(data)
      */
     async create(req, res) {
+        if(Object.keys(req.body).length === 0){
+            return res.status(404).json('Insira um valores para criar!');
+        };
+
         req.body.id = uuidv4(); // create uuid
         const data = await db.User.create(req.body) // insert body request
         res.json(data);
@@ -61,6 +63,10 @@ class UserController {
      * @returns boolean
      */
     async update(req, res) {
+        if(Object.keys(req.body).length === 0){
+            return res.status(404).json('Insira um valores para criar!');
+        };
+
         const data = await db.User.update(req.body, {where: {id:req.body.id}})
         res.json(data);
     }
@@ -72,6 +78,9 @@ class UserController {
      * @returns boolean
      */
     async destroy(req, res) {
+        if(Object.keys(req.body).length === 0){
+            return res.status(404).json('Insira um valores para criar!');
+        };
         const data = await db.User.destroy({where: {id:req.body.id}}) // update body request
         res.json(data);
     }
